Adds the NPK calculator to the lawn detail page
This adds the javascript (jquery & ajax) script which allows for no page refresh on calculation. It also adds the html to display the form. It completely changes the way the lawn_detail.html template is displayed. Instead of a view function, this creates a whole new view class with two methods. The get method contains the main display code, and the post method contains the code necessary for the NPK calculator.

closes #5